Implement remaining arguments, add README.
This finishes the main aspects of the first version of this boot integration. It
adds an extra `--options` flag to provide free form EDN options, so that plugins
can be configured from the command line as well.

It also adds `--[no]-$x` versions for `--watch` and `--color`.

It now uses a plain `System/exit` to terminate, so that no unnecessary
stacktraces are produced.

Document the general installation and usage.